// cosmowcl/src/app/shared/models/api/cosmonotifications/codiceModale.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * notifications
 * api per cosmo notifica 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CodiceModale { 
    codice?: string;
    descrizione?: string;
    codicePagina?: string;
    codiceTab?: string;
}